8月7号晚上提交

+++ internal/models/tag.go
@@ -9,7 +9,7 @@ type Tag struct {
 	Description  string `json:"description" gorm:"type:text"`        // 标签描述
 	ArticleCount int    `json:"article_count" gorm:"default:0"`      // 标签关联的文章数量
 	Heat         int    `json:"heat" gorm:"default:0"`               // 标签热度
-	FansCount    int64  `json:"fans_count" gorm:"default:0"`         // 关注人数
+	FansCount    int    `json:"fans_count" gorm:"default:0"`         // 关注人数
 	//Path         string    `json:"path" gorm:"not null"`                // 相关图片
 	Articles []Article `gorm:"many2many:article_tags;"` // 文章和标签 多对多
 	Users    []User    `gorm:"many2many:user_tags"`     // 用户和标签 多对多
